don't delete message, if media is getting edited :( Fix #2

+++ tobrot/helper_funcs/upload_to_tg.py
@@ -314,6 +314,7 @@ async def upload_single_file(message, local_file_name, caption_str, from_user, e
     except Exception as e:
         await message_for_progress_display.edit_text("**FAILED**\n" + str(e))
     else:
-        await message_for_progress_display.delete()
+        if message.message_id != message_for_progress_display.message_id:
+            await message_for_progress_display.delete()
     os.remove(local_file_name)
     return sent_message
